Move run summary helpers to appropriate modules
